+++ zoomimage-core/src/commonMain/kotlin/com/github/panpf/zoomimage/zoom/GestureType.kt
@@ -4,7 +4,6 @@ import com.github.panpf.zoomimage.annotation.IntDef
 
 @Retention(AnnotationRetention.SOURCE)
 @IntDef(
-    GestureType.NONE,
     GestureType.DRAG,
     GestureType.TWO_FINGER_SCALE,
     GestureType.ONE_FINGER_SCALE,
@@ -20,8 +19,6 @@ import com.github.panpf.zoomimage.annotation.IntDef
 annotation class GestureType {
 
     companion object {
-        const val NONE = 0
-
         const val DRAG = 1
 
         const val TWO_FINGER_SCALE = 2
@@ -30,15 +27,20 @@ annotation class GestureType {
 
         const val DOUBLE_TAP_SCALE = 8
 
-        fun name(@ContinuousTransformType type: Int): String {
+        fun name(@GestureType type: Int): String {
             return when (type) {
-                NONE -> "NONE"
                 DRAG -> "DRAG"
                 TWO_FINGER_SCALE -> "TWO_FINGER_SCALE"
                 ONE_FINGER_SCALE -> "ONE_FINGER_SCALE"
                 DOUBLE_TAP_SCALE -> "DOUBLE_TAP_SCALE"
                 else -> "UNKNOWN"
             }
+        }
+
+        val values = listOf(DRAG, TWO_FINGER_SCALE, ONE_FINGER_SCALE, DOUBLE_TAP_SCALE)
+
+        fun parse(gestureTypes: Int): List<Int> {
+            return values.asSequence().filter { gestureTypes and it != 0 }.toList()
         }
     }
 }